move tests around based on PR feedback
