+++ tests/integration/suite/daprd/scheduler/scheduler.go
@@ -14,6 +14,5 @@ limitations under the License.
 package scheduler
 
 import (
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/scheduler/grpc"
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/scheduler/http"
+	_ "github.com/dapr/dapr/tests/integration/suite/daprd/scheduler/api"
 )
