+++ tests/integration/suite/scheduler/api/http/errors.go
